Add ability to warn on use of non-official types
Consumers of os-service-types may want to warn their users about using
non-official types while still making use of the alias data to allow
things to work. Since os-service-types is at the bottom of the stack for
that, it can be useful to just configure ServiceTypes to emit the
warnings itself - that way people can be assured that all the codepaths
are hit.

Change-Id: Ia4ea392671c8196a5083ac1ab27cc1e4fb289946