Use tox_install.sh to install dependencies
Now that os-service-types exists in the requirements repository, we need
to take special care to unconstrain it when we run tests on it. That is
what the tox_install.sh script does (copied from keystoneauth in this
case). If we don't do this, tox will be unable to install test
dependencies in CI[1].

[1] http://logs.openstack.org/85/488085/4/check/gate-os-service-types-docs-ubuntu-xenial/6cb3162/console.html

Change-Id: I2726075d6f9eec33094753cda558275e433b6ca4