Consume new mapping fields from service-type-authority
The service type authority is publishing three new pre-computed
mappings. Add getters to expose them and use them where appropriate.
Also, add a method that will return all of the service data for a
project with more than one service to complement the one that only
returns the primary data.

Depends-On: Ib761cf9de875e7b80404797d4aa8d294ca56347a
Change-Id: I78d2dfa03a5a0c392a85a7a77612098b4b01f70e