Update unittests to be able to sync latest data
There have been two changes to service-types-authority data which break
the unittests for os-service-types. Some of the url schemes were set to
https instead of http, and the order of the volume aliases changed.

Update the unittests so that we can sync data updates again. Sync one by
hand to verify the updates.

Change-Id: Ic323db63b1ab8e74284bdd32a858cb1722a0d4ed
Depends-On: I3086dcfb85f5ca9970f018d503c9378f307a3e73