fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We can't set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36 due to https://github.com/tox-dev/tox/issues/425

Change-Id: Ibeb97401c13fb1cb35f824082eb6b3209c868beb