Add flag for returning unofficial types
In consuming code in openstacksdk and keystoneauth, the pattern:

  if stm.is_official(service_type):
      return stm.get_service_type(service_type)
  return service_type

Gets used in several places. Let's provide a way to say that with a
flag:

  return std.get_service_type(service_type, permissive=True)

which says "get me the official service_type for this, but if it's not
something I know about, give me the thing the user asked for.

Change-Id: Ib8fb14a88ecc690da67967b7e906deb7d923f869